+++ examples/data-sources/netapp-ontap_name_services_unix_user/data-source.tf
@@ -1,0 +1,6 @@
+data "netapp-ontap_unix_user" "unix_user" {
+  # required to know which system to interface with
+  cx_profile_name = "hw-cluster"
+  name = "root"
+  svm_name = "svm1"
+}